+++ KappaAnalysis/src/Consumers/KappaCutFlowHistogramConsumer.cc
@@ -6,7 +6,7 @@ void KappaCutFlowHistogramConsumer::Init(KappaSettings const& settings)
 {
 	CutFlowHistogramConsumer<KappaTypes>::Init(settings);
 
-	this->weightExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> float {
+	this->weightExtractor = [](event_type const& event, product_type const& product, setting_type const& setting) -> double {
 		return SafeMap::GetWithDefault(product.m_weights, std::string(setting.GetEventWeight()), 1.0);
 	};
